Move ast constructors and helpers to separate file
