+++ ast/ast.go
@@ -32,13 +32,6 @@ func (n *Scope) Type() string {
 	return SCOPE
 }
 
-func ScopeExpressions(expressions ...Expression) *Scope {
-	return &Scope{
-		Definitions: []*Definition{},
-		Expressions: expressions,
-	}
-}
-
 type Definition struct {
 	Identifier Identifier
 	Expression Expression
@@ -60,11 +53,6 @@ func (n *Identifier) Type() string {
 	return IDENTIFIER
 }
 
-func NewIdentifier(value string) *Identifier {
-	ident := Identifier(value)
-	return &ident
-}
-
 type Selector struct {
 	Items []*SelectorItem
 }
@@ -82,36 +70,12 @@ func (n *SelectorItem) Type() string {
 	return SELECTOR_ITEM
 }
 
-func NewIdentifierSelector(ident *Identifier) *SelectorItem {
-	return &SelectorItem{
-		Identifier: ident,
-	}
-}
-
-func NewIndexSelector(num *Integer) *SelectorItem {
-	return &SelectorItem{
-		Index: num,
-	}
-}
-
-func NewSelector(identifiers ...*SelectorItem) *Selector {
-	return &Selector{
-		Items: identifiers,
-	}
-}
-
 type Application struct {
 	Arguments []Expression
 }
 
 func (n *Application) Type() string {
 	return APPLICATION
-}
-
-func NewApplication(arguments ...Expression) *Application {
-	return &Application{
-		Arguments: arguments,
-	}
 }
 
 type Function struct {
@@ -150,13 +114,6 @@ func (n *Array) Type() string {
 	return ARRAY
 }
 
-func NewArray(size uint64, elements ...Expression) *Array {
-	return &Array{
-		Size:     &size,
-		Elements: elements,
-	}
-}
-
 type List struct {
 	Elements []Expression
 }
@@ -165,21 +122,10 @@ func (n *List) Type() string {
 	return LIST
 }
 
-func NewList(elements ...Expression) *List {
-	return &List{
-		Elements: elements,
-	}
-}
-
 type Integer int64
 
 func (n *Integer) Type() string {
 	return INTEGER
-}
-
-func NewInteger(value int64) *Integer {
-	i := Integer(value)
-	return &i
 }
 
 type Float float64
@@ -188,20 +134,10 @@ func (n *Float) Type() string {
 	return FLOAT
 }
 
-func NewFloat(value float64) *Float {
-	f := Float(value)
-	return &f
-}
-
 type String string
 
 func (n *String) Type() string {
 	return STRING
-}
-
-func NewString(value string) *String {
-	s := String(value)
-	return &s
 }
 
 type Keyword string
@@ -210,18 +146,8 @@ func (n *Keyword) Type() string {
 	return KEYWORD
 }
 
-func NewKeyword(value string) *Keyword {
-	k := Keyword(value)
-	return &k
-}
-
 type Boolean string
 
 func (n *Boolean) Type() string {
 	return BOOLEAN
-}
-
-func NewBoolean(value string) *Boolean {
-	b := Boolean(value)
-	return &b
 }
